bnbchain-internal/requirements_repo_gatherer.txt
requests>=2.25.0
python-dotenv>=0.19.0
pathlib2>=2.3.0; python_version < "3.4"
